refactor(hw/storage): split files
